Revert "moved GetAttorneyStatus"
This reverts commit 7f62488d74e7e1f4bb8920b505bd499cbcb8f59d.
