+++ service-api/app/src/App/src/Service/Lpa/FindActorInLpa.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Service\Lpa;
 
-use App\Service\Lpa\GetAttorneyStatus\GetAttorneyStatus;
 use Psr\Log\LoggerInterface;
 
 class FindActorInLpa
